+++ src/lib/index.js
@@ -48,7 +48,7 @@ let morning = {
         white : 'wh'
     },
     isMorning : true,
-    version : '0.12.20',
+    version : '0.12.21',
     map : {}
 };